+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use keywords to colorize menus
 - Use `gradient-string` for the titles
 - Bump versions of reactive dependencies
+- Extract all color variables to centralized object
 
 ## [3.1.0] - 2021-05-27